javascript/messages-app/requirements.txt
fastapi
mysql-connector-python
unvicorn